+++ libs/trezor-crypto/base58.h
@@ -28,7 +28,7 @@
 #include "hasher.h"
 #include "options.h"
 
-int base58_encode_check(const uint8_t *data, int len, HasherType hasher_type, char *str, int strsize);
+int base58_encode_check(const uint8_t *data, size_t len, HasherType hasher_type, char *str, size_t strsize);
 int base58_decode_check(const char *str, HasherType hasher_type, uint8_t *data, int datalen);
 
 // Private
